requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.6.3
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
Pillow==9.4.0
kaggle==1.5.13
